+++ library/PSFramework/Logging/Provider.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Management.Automation;
 using PSFramework.Utility;
@@ -40,7 +41,7 @@ namespace PSFramework.Logging
         #region Message filtering
         private List<string> _IncludeModules = new List<string>();
         /// <summary>
-        /// List of modules to include in the logging. Only messages generated from these modules will be considered by the provider
+        /// List of modules to include in the logging. Only messages generated from these modules will be considered by the provider instance.
         /// </summary>
         public List<string> IncludeModules
         {
@@ -56,7 +57,7 @@ namespace PSFramework.Logging
 
         private List<string> _ExcludeModules = new List<string>();
         /// <summary>
-        /// List of modules to exclude in the logging. Messages generated from these modules will be ignored by this provider.
+        /// List of modules to exclude in the logging. Messages generated from these modules will be ignored by this provider instance.
         /// </summary>
         public List<string> ExcludeModules
         {
@@ -70,9 +71,41 @@ namespace PSFramework.Logging
             }
         }
 
+        private List<string> _IncludeFunctions = new List<string>();
+        /// <summary>
+        /// List of functions to include in the logging. Only messages generated from these functions will be considered by the provider instance.
+        /// </summary>
+        public List<string> IncludeFunctions
+        {
+            get { return _IncludeFunctions; }
+            set
+            {
+                if (value == null)
+                    _IncludeFunctions.Clear();
+                else
+                    _IncludeFunctions = value;
+            }
+        }
+
+        private List<string> _ExcludeFunctions = new List<string>();
+        /// <summary>
+        /// List of functions to exclude in the logging. Messages generated from these functions will be ignored by this provider instance.
+        /// </summary>
+        public List<string> ExcludeFunctions
+        {
+            get { return _ExcludeFunctions; }
+            set
+            {
+                if (value == null)
+                    _ExcludeFunctions.Clear();
+                else
+                    _ExcludeFunctions = value;
+            }
+        }
+
         private List<string> _IncludeTags = new List<string>();
         /// <summary>
-        /// List of tags to include. Only messages with these tags will be considered by this provider.
+        /// List of tags to include. Only messages with these tags will be considered by this provider instance.
         /// </summary>
         public List<string> IncludeTags
         {
@@ -88,7 +121,7 @@ namespace PSFramework.Logging
 
         private List<string> _ExcludeTags = new List<string>();
         /// <summary>
-        /// List of tags to exclude. Messages with these tags will be ignored by this provider.
+        /// List of tags to exclude. Messages with these tags will be ignored by this provider instance.
         /// </summary>
         public List<string> ExcludeTags
         {
@@ -102,31 +135,100 @@ namespace PSFramework.Logging
             }
         }
 
+        private int _MinLevel = 1;
         /// <summary>
-        /// Tests whether a log entry applies to the provider
+        /// The minimum level of message to log.
+        /// Note, the lower the level, the higher the priority.
+        /// </summary>
+        public int MinLevel
+        {
+            get { return _MinLevel; }
+            set
+            {
+                if (value < 1)
+                    _MinLevel = 1;
+                else if (value > 9)
+                    _MinLevel = 9;
+                else
+                    _MinLevel = value;
+            }
+        }
+
+        private int _MaxLevel = 9;
+        /// <summary>
+        /// The maximum level of message to log.
+        /// Note, the lower the level, the higher the priority.
+        /// </summary>
+        public int MaxLevel
+        {
+            get { return _MaxLevel; }
+            set
+            {
+                if (value < 1)
+                    _MaxLevel = 1;
+                else if (value > 9)
+                    _MaxLevel = 9;
+                else
+                    _MaxLevel = value;
+            }
+        }
+
+        /// <summary>
+        /// Whether to include warning messages in the log
+        /// </summary>
+        public bool IncludeWarning = true;
+
+        /// <summary>
+        /// Tests whether a log entry applies to the provider instance
         /// </summary>
         /// <param name="Entry">The Entry to validate</param>
         /// <returns>Whether it applies</returns>
         public bool MessageApplies(Message.LogEntry Entry)
         {
-            if ((IncludeModules.Count == 0) && (ExcludeModules.Count == 0) && (IncludeTags.Count == 0) && (ExcludeTags.Count == 0))
-                return true;
+            // Level
+            if (!IncludeWarning && (Entry.Level == Message.MessageLevel.Warning))
+                return false;
+            if (((_MinLevel != 1) || (_MaxLevel != 9)) && (Entry.Level != Message.MessageLevel.Warning))
+            {
+                if (Entry.Level < (Message.MessageLevel)_MinLevel)
+                    return false;
+                if (Entry.Level > (Message.MessageLevel)_MaxLevel)
+                    return false;
+            }
 
+            // Modules
             if (IncludeModules.Count > 0)
             {
                 bool test = false;
                 foreach (string module in IncludeModules)
-                    if (string.Equals(Entry.ModuleName, module, System.StringComparison.InvariantCultureIgnoreCase))
+                    if (string.Equals(Entry.ModuleName, module, StringComparison.InvariantCultureIgnoreCase))
                         test = true;
 
                 if (!test)
                     return false;
             }
 
-            foreach(string module in ExcludeModules)
-                if (string.Equals(Entry.ModuleName, module, System.StringComparison.InvariantCultureIgnoreCase))
+            foreach (string module in ExcludeModules)
+                if (string.Equals(Entry.ModuleName, module, StringComparison.InvariantCultureIgnoreCase))
                     return false;
 
+            // Functions
+            if (IncludeFunctions.Count > 0)
+            {
+                bool test = false;
+                foreach (string function in IncludeFunctions)
+                    if (UtilityHost.IsLike(Entry.FunctionName, function))
+                        test = true;
+
+                if (!test)
+                    return false;
+            }
+
+            foreach (string function in ExcludeFunctions)
+                if (UtilityHost.IsLike(Entry.FunctionName, function))
+                    return false;
+
+            // Tags
             if (IncludeTags.Count > 0)
             {
                 if (IncludeTags.Except(Entry.Tags).ToList().Count == IncludeTags.Count)
@@ -140,20 +242,18 @@ namespace PSFramework.Logging
         }
 
         /// <summary>
-        /// Tests whether an error record applies to the provider
+        /// Tests whether an error record applies to the provider instance
         /// </summary>
         /// <param name="Record">The error record to test</param>
-        /// <returns>Whether it applies to the provider</returns>
+        /// <returns>Whether it applies to the provider instance</returns>
         public bool MessageApplies(Message.PsfExceptionRecord Record)
         {
-            if ((IncludeModules.Count == 0) && (ExcludeModules.Count == 0) && (IncludeTags.Count == 0) && (ExcludeTags.Count == 0))
-                return true;
-
+            // Modules
             if (IncludeModules.Count > 0)
             {
                 bool test = false;
                 foreach (string module in IncludeModules)
-                    if (string.Equals(Record.ModuleName, module, System.StringComparison.InvariantCultureIgnoreCase))
+                    if (string.Equals(Record.ModuleName, module, StringComparison.InvariantCultureIgnoreCase))
                         test = true;
 
                 if (!test)
@@ -161,9 +261,26 @@ namespace PSFramework.Logging
             }
 
             foreach (string module in ExcludeModules)
-                if (string.Equals(Record.ModuleName, module, System.StringComparison.InvariantCultureIgnoreCase))
+                if (string.Equals(Record.ModuleName, module, StringComparison.InvariantCultureIgnoreCase))
                     return false;
 
+            // Functions
+            if (IncludeFunctions.Count > 0)
+            {
+                bool test = false;
+                foreach (string function in IncludeFunctions)
+                    if (UtilityHost.IsLike(Record.FunctionName, function))
+                        test = true;
+
+                if (!test)
+                    return false;
+            }
+
+            foreach (string function in ExcludeFunctions)
+                if (UtilityHost.IsLike(Record.FunctionName, function))
+                    return false;
+
+            // Tags
             if (IncludeTags.Count > 0)
             {
                 if (IncludeTags.Except(Record.Tags).ToList().Count == IncludeTags.Count)
@@ -213,12 +330,12 @@ namespace PSFramework.Logging
         /// </summary>
         public void LocalizeEvents()
         {
-            UtilityHost.ImportScriptBlock(BeginEvent);
-            UtilityHost.ImportScriptBlock(StartEvent);
-            UtilityHost.ImportScriptBlock(MessageEvent);
-            UtilityHost.ImportScriptBlock(ErrorEvent);
-            UtilityHost.ImportScriptBlock(EndEvent);
-            UtilityHost.ImportScriptBlock(FinalEvent);
+            UtilityHost.ImportScriptBlock(BeginEvent, true);
+            UtilityHost.ImportScriptBlock(StartEvent, true);
+            UtilityHost.ImportScriptBlock(MessageEvent, true);
+            UtilityHost.ImportScriptBlock(ErrorEvent, true);
+            UtilityHost.ImportScriptBlock(EndEvent, true);
+            UtilityHost.ImportScriptBlock(FinalEvent, true);
         }
         #endregion Scriptblocks Logging Execution
 
@@ -226,27 +343,27 @@ namespace PSFramework.Logging
         /// <summary>
         /// Script that recognizes, whether the provider has been isntalled correctly. Some providers require installation, in order to function properly.
         /// </summary>
-        public ScriptBlock IsInstalledScript;
+        public PsfScriptBlock IsInstalledScript;
 
         /// <summary>
         /// Script that installs the provider
         /// </summary>
-        public ScriptBlock InstallationScript;
+        public PsfScriptBlock InstallationScript;
 
         /// <summary>
         /// Script that generates dynamic parameters for installing the provider.
         /// </summary>
-        public ScriptBlock InstallationParameters;
+        public PsfScriptBlock InstallationParameters;
 
         /// <summary>
         /// Scriptblock that adds additional parameters as needed to the Set-PSFLoggingProvider function
         /// </summary>
-        public ScriptBlock ConfigurationParameters;
+        public PsfScriptBlock ConfigurationParameters;
 
         /// <summary>
         /// The configuration steps taken by the Set-PSFLoggingProvider function
         /// </summary>
-        public ScriptBlock ConfigurationScript;
+        public PsfScriptBlock ConfigurationScript;
         #endregion Function Extension / Integration
 
         #region Troubleshooting / Analysis